% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_functions.R
\name{all_column_correlation}
\alias{all_column_correlation}
\title{Compare matrices and return a matrix with the correlations of all columns}
\usage{
all_column_correlation(query_mat, target_mat, method = "pearson")
}
\arguments{
\item{query_mat}{The query matrix}

\item{target_mat}{The target matrix}

\item{method}{The name of the method to pass to cor(). Can be "pearson", "kendall", or "spearman". Default is "pearson".}
}
\value{
a matrix with all correlation values, with the columns matching columns of target_mat,
and rows matching the columns of query_mat
}
\description{
For each column in the query matrix, max_column_correlation will
find the column in the target matrix with the highest correlation score,
and will return a data.frame with the correlation score and target column name.
}
