% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{balance_fragment_clusters}
\alias{balance_fragment_clusters}
\title{Downsample fragments within cluster sets to match the sample with the lowest number of fragments}
\usage{
balance_fragment_clusters(fragment_list, clusters)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects.}

\item{clusters}{a vector with cluster assignments for each item in fragment_list}
}
\value{
a lis of GenomicRanges objects with all members of each cluster downsampled to the
minimum number of reads of all cluster members.
}
\description{
Downsample fragments within cluster sets to match the sample with the lowest number of fragments
}
