% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{bam_to_fragment_list}
\alias{bam_to_fragment_list}
\title{Read multiple paired-end BAM files as a list of GRanges objects}
\usage{
bam_to_fragment_list(bamfiles)
}
\arguments{
\item{bamfiles}{A character vector of file locations for the BAM files to read}
}
\value{
a list object containing GenomicRanges objects. List names will BAM filenames without leading directories.
}
\description{
Read multiple paired-end BAM files as a list of GRanges objects
}
