% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{bed_to_GRanges}
\alias{bed_to_GRanges}
\title{Convert data.frames in BED-like format to GRanges objects}
\usage{
bed_to_GRanges(bed)
}
\arguments{
\item{bed}{A data.frame with columns chr, start, end, name, strand, and score.}
}
\value{
A GenomicRanges object
}
\description{
Convert data.frames in BED-like format to GRanges objects
}
