% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{bed_to_ucsc_loc}
\alias{bed_to_ucsc_loc}
\title{Convert GenomicRanges objects to a vector of UCSC browser locations.}
\usage{
bed_to_ucsc_loc(bed)
}
\arguments{
\item{bed}{A data.frame with columns chr, start, end, name, strand, and score.}
}
\value{
a character vector of UCSC browser locations
}
\description{
Convert GenomicRanges objects to a vector of UCSC browser locations.
}
