% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{clusterApplyLB_chunks}
\alias{clusterApplyLB_chunks}
\title{Run clusterApplyLB across a single variable, N, using chunks with user feedback.}
\usage{
clusterApplyLB_chunks(N, n_chunks = 20, cl, FUN, ...)
}
\arguments{
\item{N}{Variable passed to nodes running FUN}

\item{n_chunks}{Number of chunks to use. Default = 20 (5% of N per chunk).}

\item{cl}{A cluster object created by the parallel or snow package.}

\item{FUN}{The function to run using clusterApplyLB()}

\item{...}{Additional parameters passed to clusterApplyLB()}
}
\description{
Run clusterApplyLB across a single variable, N, using chunks with user feedback.
}
