% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{collapse_fragment_list}
\alias{collapse_fragment_list}
\title{Collapse a list of GenomicRanges objects to a single, merged GenomicRanges object}
\usage{
collapse_fragment_list(fragment_list, width = 500)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects}

\item{width}{If not NULL, whill resize the regions in each GenomicRanges object to this width before merging. Default is 500.}
}
\value{
A GenomicRanges object containing merged regions.
}
\description{
Collapse a list of GenomicRanges objects to a single, merged GenomicRanges object
}
