% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_ranges.R
\name{combine_group_GRanges}
\alias{combine_group_GRanges}
\title{Combine individual GRanges from a list of GRanges objects based on group annotations.}
\usage{
combine_group_GRanges(fragment_list, samples, group_col)
}
\arguments{
\item{fragment_list}{a list of GRanges objects, each named by sample_id}

\item{samples}{a sample meatdata data.frame with a column called sample_id and other columns
to be used for grouping.}

\item{group_col}{The column of the samples data.frame with group assignments.}
}
\value{
A list of GRanges objects, one for each unique value in group_col.
}
\description{
This is useful for making combined objects at the level of subclasses, clusters, or
other annotations for analysis within R.
}
