% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_ranges.R
\name{compare_fragment_lists}
\alias{compare_fragment_lists}
\title{Compare two lists of GRanges to get counts or overlap coverage.}
\usage{
compare_fragment_lists(query_regions, target_regions, mode = c("counts",
  "coverage"), norm = c("none", "query", "target", "total"))
}
\arguments{
\item{query_regions}{a list of GRanges objects}

\item{target_regions}{a second list of GRanges objects}

\item{mode}{Either "counts" or "coverage".}

\item{norm}{Normalization mode: "none", "query","target", or "total".}
}
\value{
a matrix with query regions as rows and target regions as columns.
}
\description{
If mode = "counts", generates a matrix for the number of overlaps between
each set of query and target GRanges. Normalization parameters specify whether
to divide by the number of regions in the query, target, or the sum of both (total).
}
\details{
If mode = "coverage", generates a matrix with the total number of overlapping bases
between each set of query and target GRanges. Normalization parameters specify whether
to divide by the total coverage of regions in the query, target, or the sum of both (total).
}
