% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{convert_fragment_list}
\alias{convert_fragment_list}
\title{Convert a list of fragments to cut sites or Tn5 footprints}
\usage{
convert_fragment_list(fragment_list, to = c("cuts", "footprints"))
}
\arguments{
\item{fragment_list}{A list of GenomicRanges objects containing fragment regions}

\item{to}{The type of regions to convert to. Can be "cuts", which are the 5' ends of each region, or "footprints",
which are the 19 bp footprint regions of the Tn5 transposase.}
}
\value{
A list of the same length of fragment_list, with objects converted to cuts or footprints.
}
\description{
Convert a list of fragments to cut sites or Tn5 footprints
}
