% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{downsample_fragments}
\alias{downsample_fragments}
\title{Equally downsample a list of GenomicRanges objects}
\usage{
downsample_fragments(fragment_list, downsample_n = 10000,
  discard_if_too_few = TRUE, seed = 42, verbose = TRUE)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects}

\item{downsample_n}{The number of fragments to retain. Default is 1e4.}

\item{discard_if_too_few}{Logical to determine if samples with fewer than downsample_n regions should be discarded.
If FALSE, samples with < downample_n regions will be retained without sampling. Default is TRUE.}

\item{seed}{A seed value for sampling for reproducibility. Default is 42.}
}
\value{
A list object containing downsampled GenomicRanges objects.
}
\description{
Equally downsample a list of GenomicRanges objects
}
