% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{expand_fragments}
\alias{expand_fragments}
\title{Expand the boundaries of regions in a list of GenomicRanges objects}
\usage{
expand_fragments(fragment_list, width = 10000, collapse = TRUE)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects}

\item{width}{The fragment size to output. Default is 1e4.}

\item{collapse}{Logical to determine if overlapping fragments should be merged after expansion. Default is TRUE.}
}
\value{
A list object containing resized GenomicRanges objects.
}
\description{
This function expands from the center of the GenomicRanges objects using GenomicRanges::resize(fix = "center")
}
