% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{fragment_overlap_jaccard_parallel_full}
\alias{fragment_overlap_jaccard_parallel_full}
\title{Linking function for computing jaccard similarity scores based on GenomicRanges overlaps in parallel.}
\usage{
fragment_overlap_jaccard_parallel_full(N)
}
\arguments{
\item{N}{Index of the comparison to compute}
}
\description{
This version returns a data.frame with extended stats for every comparison.
}
