% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{fragments_to_windows}
\alias{fragments_to_windows}
\title{Convert a list of GenomicRanges objects to a list of genomic window ids}
\usage{
fragments_to_windows(fragment_list, window_size = 10000, collapse = TRUE)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects}

\item{window_size}{The size of the genomic windows to use. Default is 1e4.}

\item{collapse}{Logical to determine if fragments found in the same window should be merged after conversion. Default is TRUE.}
}
\value{
A nested list object. Each sample in the fragment list will result in a list containing vectors with integer values for windows on each chromosome.
}
\description{
This function converts genomic positions to integer window positions based on the 5' end of each GenomicRanges region.
}
