% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_functions.R
\name{get_gene_bodies}
\alias{get_gene_bodies}
\title{Read gene body locations from the UCSC RefGenes table}
\usage{
get_gene_bodies(symbols = NULL, expand = 0, genome = "mm10")
}
\arguments{
\item{symbols}{A character object with gene symbols to retain. If NULL, will return all genes. Default is NULL.}

\item{expand}{Distance from the ends of the gene bodies to expand. If a single value is provided, will use this values in both directions.
If two values are provided, they will be used to expand regions in the 5' and 3' direction relative to transcription direction. Default is 0.}

\item{genome}{The genome to use. Default is "mm10".}
}
\value{
a data.frame with columns matching BED format (chr, start, end, name, score, and strand).
}
\description{
Read gene body locations from the UCSC RefGenes table
}
