% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_functions.R
\name{get_great_regions}
\alias{get_great_regions}
\title{Generate regions similar to GREAT based on the UCSC RefGenes table}
\usage{
get_great_regions(symbols = NULL, minexpand = 5000, maxexpand = 1e+06,
  genome = "mm10")
}
\arguments{
\item{symbols}{A character object with gene symbols to retain. If NULL, will return all genes. Default is NULL.}

\item{minexpand}{Minimum distance from the TSS to expand. This region will be retained even if it overlaps an adjacent gene. Default is 5000.}

\item{genome}{The genome to use. Default is "mm10".}

\item{minexpand}{Maximum distance from the TSS to expand. This region will be bounded by adjacent minimum regions of the nearest gene.}
}
\value{
a data.frame with columns matching BED format (chr, start, end, name, score, and strand).
}
\description{
GREAT is the [Genomic Regions Enrichment of Annotations Tool](http://great.stanford.edu).
}
