% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{GRanges_to_bed}
\alias{GRanges_to_bed}
\title{Convert UCSC-style genomic locations to a BED-like data.frame}
\usage{
GRanges_to_bed(gr)
}
\arguments{
\item{gr}{A GenomicRanges object}
}
\value{
a data.frame with chr, start, end, and strand.
}
\description{
Convert UCSC-style genomic locations to a BED-like data.frame
}
