% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{GRanges_to_ucsc_loc}
\alias{GRanges_to_ucsc_loc}
\title{Convert GenomicRanges objects to a vector of UCSC browser locations.}
\usage{
GRanges_to_ucsc_loc(gr)
}
\arguments{
\item{gr}{A GenomicRanges object}
}
\value{
a character vector of UCSC browser locations
}
\description{
Convert GenomicRanges objects to a vector of UCSC browser locations.
}
