% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{jaccard_df_to_mat}
\alias{jaccard_df_to_mat}
\title{Convert jaccard similarity results from data.frame-based functions to a matrix}
\usage{
jaccard_df_to_mat(res)
}
\arguments{
\item{res}{The results from overlap_jaccard_df() or window_jaccard_df()}
}
\value{
a jaccard distance matrix
}
\description{
Convert jaccard similarity results from data.frame-based functions to a matrix
}
