% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_functions.R
\name{matrix_var_filter}
\alias{matrix_var_filter}
\title{Filter a matrix based on standard deviation of variance}
\usage{
matrix_var_filter(x, sd_cut = 2, top = NULL)
}
\arguments{
\item{x}{The matrix to filter}

\item{sd_cut}{The cutoff for sd(var()). Default = 2. If NULL, performs no filtering.}

\item{top}{The max number of rows to return. If NULL, returns all results.}
}
\value{
a matrix with filtered rows.
}
\description{
This filters the rows of a matrix by computing the log of variance + 1 for each row,
computing the standard deviation of all log_vars, and then selecting rows
with a log_var greater than the value of mean(log_var) + sd(log_var) * sd_cut.
}
