% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_functions.R
\name{max_column_correlation}
\alias{max_column_correlation}
\title{Compare matrices to find the maximum correlation of columns}
\usage{
max_column_correlation(query_mat, target_mat, method = "pearson")
}
\arguments{
\item{query_mat}{The query matrix}

\item{target_mat}{The target matrix}

\item{method}{The name of the method to pass to cor(). Can be "pearson", "kendall", or "spearman". Default is "pearson".}
}
\value{
a data.frame with 3 columns: query, max_cor, and target.
}
\description{
For each column in the query matrix, max_column_correlation will
find the column in the target matrix with the highest correlation score,
and will return a data.frame with the correlation score and target column name.
}
