% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{merge_fragment_list}
\alias{merge_fragment_list}
\title{Merge a list of GenomicRanges objects to a single, merged GenomicRanges object}
\usage{
merge_fragment_list(fragment_list)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects}
}
\value{
A GenomicRanges object containing merged regions.
}
\description{
Merge a list of GenomicRanges objects to a single, merged GenomicRanges object
}
