% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{overlap_jaccard_mat}
\alias{overlap_jaccard_mat}
\title{Compute jaccard similarity values based on GenomicRanges overlaps in parallel.}
\usage{
overlap_jaccard_mat(fragment_list, n_cores = "auto",
  cluster_type = "PSOCK")
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects.}

\item{n_cores}{The number of cores to use in parallel. Use "auto" to detect and use all cores. Default is 6.}

\item{cluster_type}{Either "PSOCK" (for Windows) or "FORK" (possible on Linux and Mac). FORK is more memory-efficient.}
}
\value{
a matrix of jaccard distances
}
\description{
Compute jaccard similarity values based on GenomicRanges overlaps in parallel.
}
