% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{overlap_jaccard_mat_fun}
\alias{overlap_jaccard_mat_fun}
\title{Linking function for computing jaccard similarity scores based on GenomicRanges overlaps in parallel.}
\usage{
overlap_jaccard_mat_fun(N)
}
\arguments{
\item{N}{Index of the comparison to compute}
}
\description{
Linking function for computing jaccard similarity scores based on GenomicRanges overlaps in parallel.
}
