% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{pe_to_frag}
\alias{pe_to_frag}
\title{Read a paired-end BAM file as GRanges fragments}
\usage{
pe_to_frag(bamfile)
}
\arguments{
\item{bamfile}{The BAM file to read}
}
\value{
a GenomicRanges object
}
\description{
Read a paired-end BAM file as GRanges fragments
}
