% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{pe_to_frag_parallel}
\alias{pe_to_frag_parallel}
\title{Linking function for running pe_to_frag in parallel mode.}
\usage{
pe_to_frag_parallel(N)
}
\arguments{
\item{N}{Index of the bam_file to read}
}
\description{
Linking function for running pe_to_frag in parallel mode.
}
