% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{percent_display}
\alias{percent_display}
\title{Convert i and n_chunks into a percentage character object.}
\usage{
percent_display(i, n_chunks)
}
\arguments{
\item{i}{The current value}

\item{n_chunks}{number of chunks}
}
\value{
a character object.
}
\description{
Convert i and n_chunks into a percentage character object.
}
