% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_overlap_functions.R
\name{region_fragment_count_clusters}
\alias{region_fragment_count_clusters}
\title{Aggregate sample overlaps by cluster}
\usage{
region_fragment_count_clusters(fragment_list, fragment_clusters, regions,
  name_col = "mcols.name")
}
\arguments{
\item{fragment_list}{The list of GenomicRanges objects to compare to regions}

\item{regions}{A single GenomicRanges object as targets}

\item{name_col}{The column in regions that has the name of each region.}
}
\value{
A matrix with each fragment list as columns and each target region as rows containing
the counts of overlaps between each fragment list in each region.
}
\description{
Aggregate sample overlaps by cluster
}
