% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{res_to_distance_matrix}
\alias{res_to_distance_matrix}
\title{Convert parallel jaccard similarity results from data.frame to a matrix}
\usage{
res_to_distance_matrix(res)
}
\arguments{
\item{res}{The results from run_fragment_overlap_jaccard_parallel() or run_window_overlap_jaccard_parallel()}
}
\value{
a jaccard distance matrix
}
\description{
Convert parallel jaccard similarity results from data.frame to a matrix
}
