% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{run_fragment_overlap_jaccard_parallel}
\alias{run_fragment_overlap_jaccard_parallel}
\title{Compute jaccard similarity values based on GenomicRanges overlaps in parallel.}
\usage{
run_fragment_overlap_jaccard_parallel(fragment_list, n_cores = 6)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects.}

\item{n_cores}{The number of cores to use in parallel. Use "auto" to detect and use all cores. Default is 6.}

\item{sample_names}{Sample names. If NULL, will use BAM file names.}
}
\value{
a list of GenomicRanges objects
}
\description{
Compute jaccard similarity values based on GenomicRanges overlaps in parallel.
}
