% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{run_window_overlap_jaccard_parallel}
\alias{run_window_overlap_jaccard_parallel}
\title{For compatibility with older scripts}
\usage{
run_window_overlap_jaccard_parallel(window_list, n_cores = "auto",
  cluster_type = "PSOCK")
}
\description{
See ?window_jaccard_df for current function.
}
