% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{se_to_cuts}
\alias{se_to_cuts}
\title{Read a single-end BAM file as GRanges cut sites based on the 5' end of each read}
\usage{
se_to_cuts(bamfile)
}
\arguments{
\item{bamfile}{The BAM file to read}
}
\value{
a GenomicRanges object
}
\description{
Read a single-end BAM file as GRanges cut sites based on the 5' end of each read
}
