% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{se_to_cuts_parallel}
\alias{se_to_cuts_parallel}
\title{Linking function for running se_to_cuts in parallel mode.}
\usage{
se_to_cuts_parallel(N)
}
\arguments{
\item{N}{Index of the bam_file to read}
}
\description{
Linking function for running se_to_cuts in parallel mode.
}
