% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelization.R
\name{time_display}
\alias{time_display}
\title{Display time elapsed and projected completion time.}
\usage{
time_display(i, n_chunks, start_time)
}
\arguments{
\item{i}{The current chunk under analysis}

\item{n_chunks}{number of total chunks}

\item{start_time}{The start time of the first chunk}
}
\value{
a character object.
}
\description{
Display time elapsed and projected completion time.
}
