% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_functions.R
\name{ucsc_loc_to_bed}
\alias{ucsc_loc_to_bed}
\title{Convert GenomicRanges to a bed-like data.frame}
\usage{
ucsc_loc_to_bed(ucsc_loc)
}
\arguments{
\item{ucsc_loc}{A vector of UCSC-like genomic locations. Example: "chr1:152,548,974-152,550,854"}
}
\value{
a BED-like data.frame with chr, start, end, and strand (+)
}
\description{
Convert GenomicRanges to a bed-like data.frame
}
