% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc_functions.R
\name{ucsc_loc_to_GRanges}
\alias{ucsc_loc_to_GRanges}
\title{Convert UCSC-style genomic locations to a GenomicRegions object.}
\usage{
ucsc_loc_to_GRanges(ucsc_loc)
}
\arguments{
\item{ucsc_loc}{A vector of UCSC-like genomic locations. Example: "chr1:152,548,974-152,550,854"}
}
\value{
a GenomicRanges object for the UCSC locations
}
\description{
Convert UCSC-style genomic locations to a GenomicRegions object.
}
