% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{window_jaccard_df}
\alias{window_jaccard_df}
\title{Compute jaccard similarity values based on window overlaps in parallel.}
\usage{
window_jaccard_df(window_list, n_cores = "auto",
  cluster_type = "PSOCK")
}
\arguments{
\item{n_cores}{The number of cores to use in parallel. Use "auto" to detect and use all cores. Default is 6.}

\item{cluster_type}{Either "PSOCK" (for Windows) or "FORK" (possible on Linux and Mac). FORK is more memory-efficient.}

\item{fragment_list}{The list object containing genomic windows.}
}
\value{
a list of GenomicRanges objects
}
\description{
Compute jaccard similarity values based on window overlaps in parallel.
}
