% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{window_jaccard_df_fun}
\alias{window_jaccard_df_fun}
\title{Linking function for computing jaccard similarity scores based on window overlaps in parallel.}
\usage{
window_jaccard_df_fun(N)
}
\arguments{
\item{N}{Index of the comparison to compute}
}
\description{
Linking function for computing jaccard similarity scores based on window overlaps in parallel.
}
