% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{window_overlap_jaccard_parallel}
\alias{window_overlap_jaccard_parallel}
\title{# Returns only jd
This version is slower.}
\usage{
window_overlap_jaccard_parallel(N)
}
\arguments{
\item{N}{Index of the comparison to compute}
}
\description{
# Returns only jd
This version is slower.
}
