% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{count_fragment_overlaps}
\alias{count_fragment_overlaps}
\title{Count the number of times a list of GenomicRanges objects overlaps another list of GenomicRanges objects.}
\usage{
count_fragment_overlaps(fragment_list, target_GRanges, binarize = TRUE,
  sparse = FALSE, aggregate = FALSE)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects.}

\item{target_GRanges}{A second list of GenomicRanges objects.}

\item{binarize}{Logical indicating whether to retain count values for each region on count overlaps as binary. Default is TRUE.}

\item{sparse}{Logical indicating if the results matrix should be a sparse dgCMatrix. Only used if aggregate is FALSE. Default is FALSE.}

\item{aggregate}{Logical indicating whether to return a matrix of counts or a vector of count sums. Default is FALSE.}
}
\value{
If aggregate == TRUE, a vector with the total count of overlaps between each object in fragment_list and all objects in target_GRanges.

If aggregate == FALSE & sparse = FALSE, a matrix with count values for overlaps between each object in fragment_list (as columns) and each object in target_GRanges (as rows).

if aggregate == FALSE & sparse == TRUE, a dgCMatrix with count values for overlaps with each fragment_list as columns and each object in the target_GRanges as rows.
}
\description{
Count the number of times a list of GenomicRanges objects overlaps another list of GenomicRanges objects.
}
