% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{count_fragment_overlaps}
\alias{count_fragment_overlaps}
\title{Count the number of times a list of GenomicRanges objects overlaps another list of GenomicRanges objects.}
\usage{
count_fragment_overlaps(fragment_list, target_GRanges, binarize = TRUE,
  aggregate = FALSE)
}
\arguments{
\item{fragment_list}{The list object containing GenomicRanges objects.}

\item{target_GRanges}{A second list of GenomicRanges objects.}

\item{binarize}{Logical indicating whether to retain count values for each region on count overlaps as binary. Default is TRUE.}

\item{aggregate}{Logical indicating whether to return a matrix of counts or a vector of count sums.}
}
\value{
If aggregate == TRUE, a vector with the total count of overlaps between each object in fragment_list and all objects in target_GRanges.
If aggregate == FALSE, a matrix with count values for overlaps between each object in fragment_list (as columns) and each object in target_GRanges (as rows).
}
\description{
Count the number of times a list of GenomicRanges objects overlaps another list of GenomicRanges objects.
}
