% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_and_formats.R
\name{merge_bam_files}
\alias{merge_bam_files}
\title{Merge multiple BAM files to a single output BAM file}
\usage{
merge_bam_files(bam_files, out_file, index_out_file = TRUE)
}
\arguments{
\item{bam_files}{a character vector listing all of the BAM files to merge}

\item{out_file}{the target BAM file to write to}

\item{index_out_file}{whether or not to index the output file. Default is TRUE.}
}
\description{
Merge multiple BAM files to a single output BAM file
}
