% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pileup_gr_list}
\alias{pileup_gr_list}
\title{Pileup reads, fragments, or cuts from a BAM file over target GRanges regions.}
\usage{
pileup_gr_list(gr_list, gr_target, gr_groups = NULL, norm = "PM",
  window_size = NULL, window_mode = c("max", "mean", "median"))
}
\arguments{
\item{window_mode}{}
}
\description{
Pileup reads, fragments, or cuts from a BAM file over target GRanges regions.
}
