% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{run_fragment_overlap_jaccard_parallel_full}
\alias{run_fragment_overlap_jaccard_parallel_full}
\title{For compatibility with older scripts}
\usage{
run_fragment_overlap_jaccard_parallel_full(fragment_list,
  n_cores = "auto", cluster_type = "PSOCK")
}
\description{
See ?overlap_jaccard_df for current function.
}
